Added loader script.
The loader script contains an init function that loads the Googlemaps
api and all the libraries.

This should really only be seen as a stop-gap solution. The main problem
with this (apart from forcing the user to call an init function) is that
if the user resets the kernel and re-calls the init function, the JS
will be injected into the DOM more than once.