Rewrote heatmap_view as a JS module.
This means it can now be installed as a nbextension. By putting the call
to load Google maps in there as well, we remove the init statement
completely. However, this requires the user running

%%javascript

IPython.load_extensions("js/heatmap_view")

which is not ideal, obviously. But I think this can be re-written as a
line magic.