Added plainmap function to just draw a map.
Can now use gmaps.plainmap() to draw a Google map with no overlay.
The rationale is that it will be easier to develop the documentation
pipeline using this as a test model, rather than the more complex
heatmap.

All the views are in a single js file at the moment. This makes sense
for now, but it might be worth thinking about refactoring this in the
future.

The Python models will also need refactoring.