Switch to ECMA 2015 version of JS
By providing arrow functions, this will save a lot of rebinding that ->
this and other JS mishaps. The intention is to provide a smoother
development experience.

This commit will have no effect on users. It will not affect browser
compatibility either since we use babel and webpack to transpile to a
well accepted javascript standard anyway.