Fix bug where the initial map bounds were incorrect
Prior to this commit, the longitudinal bounds on data for heatmaps was
incorrectly calculated. This is trickier than it appears at first: since
longitudes are effectively distributed on a circle (i.e. l + 360 == l for any
longitude l), the concept of bounds is not even very well defined. Thus, just
taking the minimum and maximum longitdues as they appear in the data -- what
was done prior to this commit -- does not really make sense.

Nevertheless, in many cases the data is centred around some particular
longitudes. It thus makes sense to center the map around these longitudes. This
commit estimates the extent by looking at the variance assuming the samples are
distributed according to a wrapped normal distribution.

It also changes how bounds on latitudes (for heatmaps) are calculated: the
bounds are now 2 standard deviations around the mean latitude.

The Direction layer is left unchanged. This is unlikely to be a problem since there are
typically far fewer points in directions than in heatmaps, and they are likely to be quite
close to each other.