Remove error thrown by layer status change
This error wasn't catchable anyway, so it had no value as an exception. Its
only value was signalling to the user that something had gone wrong. This
is now handled in the frontend, so this was unnecessary.

I will keep layer_status for now, since there may be users that depend on it.